Experimentally introduced perlinNoise3d.h for adding vertical randomness to soil type distribution. Also made grass cells only appear near the surface.

+++ perlinNoise.h
@@ -114,7 +114,7 @@ void perlin_noise(const PerlinNoiseParams &param, PerlinNoiseCallback &callback)
 
 	// Return result
 	for(int xi = 0; xi < CELLSIZE; xi++) for(int yi = 0; yi < CELLSIZE; yi++){
-		callback((double)work2[xi][yi] / baseMax / sumfactor, xi, yi);
+		callback(float((double)work2[xi][yi] / baseMax / sumfactor), xi, yi);
 	}
 }
